--- src/main/java/edu/bsu/cs455/calculator/Calculator.java
@@ -1,4 +1,0 @@
-package edu.bsu.cs455.calculator;
-
-public class Calculator {
-}
